@@ -1,10 +1,10 @@
 package com.roskart.dropwizard.jaxws.example.ws;
 
-import javax.jws.WebMethod;
-import javax.jws.WebParam;
-import javax.jws.WebResult;
-import javax.jws.WebService;
-import javax.xml.ws.WebFault;
+import jakarta.jws.WebMethod;
+import jakarta.jws.WebParam;
+import jakarta.jws.WebResult;
+import jakarta.jws.WebService;
+import jakarta.xml.ws.WebFault;
 
 @WebService(targetNamespace = "http://com.roskart.dropwizard.example/JavaFirstService")
 public interface JavaFirstService {
