@@ -2,8 +2,8 @@ package com.roskart.dropwizard.jaxws.example.ws;
 
 import com.codahale.metrics.annotation.Metered;
 
-import javax.jws.WebMethod;
-import javax.jws.WebService;
+import jakarta.jws.WebMethod;
+import jakarta.jws.WebService;
 
 @WebService
 public class SimpleService {
