@@ -4,7 +4,7 @@ import com.google.common.collect.ImmutableList;
 import org.apache.cxf.interceptor.Interceptor;
 import org.apache.cxf.message.Message;
 
-import javax.xml.ws.handler.Handler;
+import jakarta.xml.ws.handler.Handler;
 
 import static com.google.common.base.Preconditions.checkArgument;
 
