@@ -1,7 +1,7 @@
 package com.roskart.dropwizard.jaxws;
 
-import javax.jws.WebMethod;
-import javax.jws.WebService;
+import jakarta.jws.WebMethod;
+import jakarta.jws.WebService;
 
 @WebService
 public class DummyService {
